Move `ioctl` up to `device_channel`
